Use more traditional login flow
Ditch logging in via AJAX form, was overly complex.

+++ project/views/gallery.tpl
@@ -1,3 +1,4 @@
+{% from "utils.tpl" import login_status with context %}
 {% macro render_pagination(pagination) %}
   {% if pagination.has_prev %}
     <a class="prev" href="{{ url_for_page(pagination.page - 1) }}" data-page="{{ page }}">&laquo;</a>
@@ -153,19 +154,22 @@
           </div>
           {% if not xhr %}
         </div>
-        <div class="row delete-options">
-          {{ form_delete.csrf_token }}
-          {{ form_delete.key(value=key) }}
-          <small>{{ form_delete.password.label }}</small> {{ form_delete.password(autocomplete="current-password") }}
-          {{ form_delete.delete_selected() }}
-          {{ form_delete.delete_all() }}
-        </div>
+        {% if current_user.is_authenticated and current_user.key == key %}
+          <div class="row delete-options">
+            {{ form_delete.csrf_token }}
+            <small>{{ login_status(show_button=False) }}</small>
+            {{ form_delete.delete_selected() }}
+            {{ form_delete.delete_all() }}
+          </div>
+        {% endif %}
       </form>
-      <div class="row padded">
-        <form action="/gallery/delete/advanced" method="get" class="main_form">
-          <input type="submit" value="Advanced Delete..." />
-        </form>
-      </div>
+      {% if current_user.is_authenticated and current_user.key == key %}
+        <div class="row padded">
+          <form action="/gallery/delete/advanced" method="get" class="main_form">
+            <input type="submit" value="Advanced Delete..." />
+          </form>
+        </div>
+      {% endif %}
     </div>
   </div>
   <script type="text/javascript">
